refactor(alerts): address code-review feedback on similarity-alert cron
Reviewer blockers:
1. Document service-role implicit bypass of compute_spot_similarity_score's
   auth.uid() guard. The RPC's `IF p_user_id <> auth.uid() THEN RAISE`
   does not raise under service-role because auth.uid() returns NULL and
   `<uuid> <> NULL` evaluates to NULL. Cron relies on this to impersonate
   each user; documented at the call site so a future RPC guard tightening
   doesn't silently break the cron.

2. Document that getUserEntitlement is a stub returning "free" unless
   ALERT_PREVIEW_MODE=true. Until RevenueCat → user_entitlements mirror
   lands (paywall spec), similarity_alert rules can't be created in
   non-preview prod, so this cron is a no-op. Deployment contract
   comment spelled out in the route header.

Should-fix:
3. NaN-threshold silent-fail guard. typeof NaN === "number" would let a
   malformed similarity_threshold through unchanged, making score < NaN
   false for every hour and the rule never fire. Extracted
   resolveSimilarityThreshold() in presets.ts using Number.isFinite,
   unit-tested against NaN, ±Infinity, strings, null/undefined, and
   missing-conditions.

4. Extract consolidateMatchedHours from the Next.js route module into
   lib/alerts/consolidate.ts so Jest tests don't have to evaluate
   NextResponse + createSupabaseServiceRoleClient top-level to test a
   pure function. Test file moved __tests__/api/cron/... →
   __tests__/lib/alerts/consolidate.test.ts.

5. Tighten the 65-minute contiguity gap's comment — forecasts are on a
   UTC grid so DST never creates odd gaps; the 5-minute buffer is for
   clock skew, not DST.

6. Document earliest-timestamp-wins tiebreaker on peak-score reduce.

92 alerts tests pass (6 new threshold-resolution cases). Typecheck clean.
