+++ __tests__/lib/alerts/presets.test.ts
@@ -3,6 +3,7 @@ import {
   getPreset,
   getPresetsForGroup,
   SIMILARITY_ALERT_DEFAULT_THRESHOLD,
+  resolveSimilarityThreshold,
 } from "@/lib/alerts/presets";
 import type { BeachAlertMeta } from "@/lib/alerts/types";
 
@@ -100,5 +101,61 @@ describe("presets", () => {
   it("SIMILARITY_ALERT_DEFAULT_THRESHOLD is in the RPC's 0–10 range", () => {
     expect(SIMILARITY_ALERT_DEFAULT_THRESHOLD).toBeGreaterThanOrEqual(0);
     expect(SIMILARITY_ALERT_DEFAULT_THRESHOLD).toBeLessThanOrEqual(10);
+  });
+});
+
+describe("resolveSimilarityThreshold", () => {
+  it("returns the stored threshold when finite", () => {
+    expect(resolveSimilarityThreshold({ similarity_threshold: 8.2 })).toBe(8.2);
+  });
+
+  it("returns 0 when stored threshold is exactly 0 (not falsy-fallback)", () => {
+    // A user might set threshold=0 to receive every computable match. Don't
+    // let a truthy-check silently overwrite that with the default.
+    expect(resolveSimilarityThreshold({ similarity_threshold: 0 })).toBe(0);
+  });
+
+  it("falls back to default for NaN", () => {
+    // typeof NaN === "number" would pass a naive guard; NaN threshold would
+    // make `score < NaN` false for every hour and the rule would never fire.
+    expect(
+      resolveSimilarityThreshold({ similarity_threshold: NaN }),
+    ).toBe(SIMILARITY_ALERT_DEFAULT_THRESHOLD);
+  });
+
+  it("falls back to default for Infinity and -Infinity", () => {
+    expect(
+      resolveSimilarityThreshold({ similarity_threshold: Infinity }),
+    ).toBe(SIMILARITY_ALERT_DEFAULT_THRESHOLD);
+    expect(
+      resolveSimilarityThreshold({ similarity_threshold: -Infinity }),
+    ).toBe(SIMILARITY_ALERT_DEFAULT_THRESHOLD);
+  });
+
+  it("falls back to default for non-number types (string, null, undefined, object)", () => {
+    expect(
+      resolveSimilarityThreshold({ similarity_threshold: "7.5" as any }),
+    ).toBe(SIMILARITY_ALERT_DEFAULT_THRESHOLD);
+    expect(
+      resolveSimilarityThreshold({ similarity_threshold: null as any }),
+    ).toBe(SIMILARITY_ALERT_DEFAULT_THRESHOLD);
+    expect(
+      resolveSimilarityThreshold({ similarity_threshold: undefined }),
+    ).toBe(SIMILARITY_ALERT_DEFAULT_THRESHOLD);
+    expect(
+      resolveSimilarityThreshold({ similarity_threshold: {} as any }),
+    ).toBe(SIMILARITY_ALERT_DEFAULT_THRESHOLD);
+  });
+
+  it("accepts null/undefined conditions object", () => {
+    expect(resolveSimilarityThreshold(null)).toBe(
+      SIMILARITY_ALERT_DEFAULT_THRESHOLD,
+    );
+    expect(resolveSimilarityThreshold(undefined)).toBe(
+      SIMILARITY_ALERT_DEFAULT_THRESHOLD,
+    );
+    expect(resolveSimilarityThreshold({})).toBe(
+      SIMILARITY_ALERT_DEFAULT_THRESHOLD,
+    );
   });
 });
